requirements copy.txt
aniso8601==9.0.1
anyio==3.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
fastapi==0.95.1
filelock==3.12.2
Flask==2.2.3
Flask-RESTful==0.3.9
fsspec==2023.6.0
gradio_client==0.2.7
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
neo4j==5.7.0
packaging==23.1
pycparser==2.21
pydantic==1.10.7
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
requests==2.28.2
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
stripe==6.6.0
tqdm==4.66.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
websockets==11.0.3
Werkzeug==2.2.3
